requirements.txt
django==4.2.7
jq==1.6.0
langchain==0.0.340
openai==1.3.5
pypdf==3.17.1
vectordb2==0.1.9